Week 1/Feature Based Image Alignment/streamlit interface/requirements.txt
numpy==1.23.2
Pillow==9.2.0
requests==2.28.1
streamlit==1.12.0